training-ia/requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
Jinja2==3.1.3
kiwisolver==1.4.7
lightning-utilities==0.11.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pillow==10.2.0
pycocotools==2.0.8
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchmetrics==1.4.2
torchvision==0.19.1+cu118
typing_extensions==4.9.0
